Refactor move ModalDialog to 'gui.base' namespace
This class neds to be extracted from widgets.py before
any modal dialog can be.
